requirements.txt
fastapi==0.121.0
starlette==0.49.3
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.24.1
langchain==1.0.5
langchain-openai==1.0.2
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-chroma
openai==1.109.1
python-dotenv==1.0.0
chromadb==1.3.4